Move signer codec to dialog package
